+++ src/CleanRegex/Internal/Replace/Counting/CallbackCountingStrategy.php
@@ -1,0 +1,19 @@
+<?php
+namespace TRegx\CleanRegex\Internal\Replace\Counting;
+
+class CallbackCountingStrategy implements CountingStrategy
+{
+    /** @var callback */
+    private $countReceiver;
+
+    public function __construct(callable $countReceiver)
+    {
+        $this->countReceiver = $countReceiver;
+    }
+
+    public function count(int $replaced): void
+    {
+        $receiver = $this->countReceiver;
+        $receiver($replaced);
+    }
+}