+++ src/CleanRegex/Internal/Replace/Counting/CountingStrategy.php
@@ -1,0 +1,7 @@
+<?php
+namespace TRegx\CleanRegex\Internal\Replace\Counting;
+
+interface CountingStrategy
+{
+    public function count(int $replaced): void;
+}