+++ src/CleanRegex/Internal/Replace/Counting/IgnoreCounting.php
@@ -1,0 +1,9 @@
+<?php
+namespace TRegx\CleanRegex\Internal\Replace\Counting;
+
+class IgnoreCounting implements CountingStrategy
+{
+    public function count(int $replaced): void
+    {
+    }
+}