+++ src/CleanRegex/Replace/Callback/ReplacePatternCallbackInvoker.php
@@ -4,6 +4,7 @@ namespace TRegx\CleanRegex\Replace\Callback;
 use TRegx\CleanRegex\Internal\InternalPattern as Pattern;
 use TRegx\CleanRegex\Internal\Model\Matches\RawMatchesOffset;
 use TRegx\CleanRegex\Internal\Replace\By\NonReplaced\SubjectRs;
+use TRegx\CleanRegex\Internal\Replace\Counting\CountingStrategy;
 use TRegx\CleanRegex\Internal\Subjectable;
 use TRegx\SafeRegex\preg;
 
@@ -17,18 +18,22 @@ class ReplacePatternCallbackInvoker
     private $limit;
     /** @var SubjectRs */
     private $substitute;
+    /** @var CountingStrategy */
+    private $countingStrategy;
 
-    public function __construct(Pattern $pattern, Subjectable $subject, int $limit, SubjectRs $substitute)
+    public function __construct(Pattern $pattern, Subjectable $subject, int $limit, SubjectRs $substitute, CountingStrategy $countingStrategy)
     {
         $this->pattern = $pattern;
         $this->subject = $subject;
         $this->limit = $limit;
         $this->substitute = $substitute;
+        $this->countingStrategy = $countingStrategy;
     }
 
     public function invoke(callable $callback, ReplaceCallbackArgumentStrategy $strategy): string
     {
         $result = $this->pregReplaceCallback($callback, $replaced, $strategy);
+        $this->countingStrategy->count($replaced);
         if ($replaced === 0) {
             return $this->substitute->substitute($this->subject->getSubject()) ?? $result;
         }