+++ test/Unit/CleanRegex/Internal/PatternLimitTest.php
@@ -7,6 +7,7 @@ use Test\Utils\Functions;
 use TRegx\CleanRegex\Internal\InternalPattern;
 use TRegx\CleanRegex\Internal\PatternLimit;
 use TRegx\CleanRegex\Internal\Replace\By\NonReplaced\DefaultStrategy;
+use TRegx\CleanRegex\Internal\Replace\Counting\IgnoreCounting;
 use TRegx\CleanRegex\Remove\RemoveLimit;
 use TRegx\CleanRegex\Replace\ReplaceLimitImpl;
 use TRegx\CleanRegex\Replace\ReplacePatternImpl;
@@ -60,7 +61,7 @@ class PatternLimitTest extends TestCase
             [
                 new ReplaceLimitImpl(function (int $limit) {
                     return new ReplacePatternImpl(
-                        new SpecificReplacePatternImpl(InternalPattern::pcre('//'), '', $limit, new DefaultStrategy()), InternalPattern::pcre('//'), '', $limit
+                        new SpecificReplacePatternImpl(InternalPattern::pcre('//'), '', $limit, new DefaultStrategy(), new IgnoreCounting()), InternalPattern::pcre('//'), '', $limit
                     );
                 })
             ],